feat: start server with cluster by default
